refactor: use '-Tab' in dir and component names instead of section
for clarity
