+++ packages/web/src/views/Calendar/components/Sidebar/MonthTab/index.ts
@@ -1,0 +1,1 @@
+export { MonthTab as ToggleableMonthWidget } from "./MonthTab";
