+++ packages/web/src/views/Calendar/components/Sidebar/SomedayTab/SomedayEvents/SomedayEventRectangle.tsx
