--- packages/web/src/views/Calendar/components/Sidebar/ToggleableMonthWidget/index.ts
@@ -1,1 +1,0 @@
-export { ToggleableMonthWidget } from "./ToggleableMonthWidget";
